misc/HtmlTestSuite/results/test-0733.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>A zero-width new formatting context should fit in a zero-width layout opportunity between floats, above a 100% wide float</title>
<link href="mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://www.w3.org/TR/CSS22/visuren.html#float-position" rel="help" title="9.5.1 Positioning the float: the 'float' property"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
</head><body><div id="container" style="position:relative; width:200px;">
  <div style="float:left; width:100px; height:200px;"/>
  <div style="float:right; width:100px; height:200px;"/>
  <div style="float:right; width:100%; height:10px;"/>
  <div data-offset-x="100" data-offset-y="0" style="overflow:hidden; width:0; height:200px;"/>
</div>
<script>
  checkLayout("#container");
</script>
</body></html>
